ability-voice/pyproject.toml
[project]
name = "ability-voice"
version = "1.0.0"
description = "Voice operations ability - TTS (Piper), STT (Whisper), wake word detection"
requires-python = ">=3.10"
dependencies = [
    "kadi @ git+https://gitlab.com/humin-game-lab/kadi/kadi-core-py.git@v0.9.0",
    "numpy",
    "sounddevice",
    "soundfile",
    "piper-tts",
    "openai-whisper",
    "webrtcvad",
    "pydantic>=2.0",
    "python-dotenv",
]

[project.optional-dependencies]
cuda = ["torch>=2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ability_voice"]
